Trung

+++ src/main/java/com/example/finallaptrinhweb/model/Product.java
@@ -8,7 +8,7 @@ public class Product {
     private String productCode;
     private String productName;
     private int categoryId;
-    private BigDecimal price;
+    private double price;
     private BigDecimal discountPrice;
     private int quantity;
     private int stockQuantity;
@@ -24,7 +24,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(int id, String productCode, String productName, int categoryId, BigDecimal price, BigDecimal discountPrice, int quantity, int stockQuantity, String ingredients, String dosage, String instructions, int warrantyPeriod, String productType, int supplierId, String imageUrl, boolean active) {
+    public Product(int id, String productCode, String productName, int categoryId, double price, BigDecimal discountPrice, int quantity, int stockQuantity, String ingredients, String dosage, String instructions, int warrantyPeriod, String productType, int supplierId, String imageUrl, boolean active) {
         this.id = id;
         this.productCode = productCode;
         this.productName = productName;
@@ -75,11 +75,11 @@ public class Product {
         this.categoryId = categoryId;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return this.price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
